\name{gDist}
\alias{gDist}
\title{Density/Distribution Plot for gPCA}
\description{
This function produces a density plot of the permutation \eqn{\delta_p} values.
}
\usage{
gDist(out)
}
\arguments{
  \item{out}{object resulting from \code{gPCA.batchdetect()} call.}
}
\references{
Reese, S. E., Archer, K. J., Therneau, T. M., Atkinson, E. J., Vachon, C. M., de Andrade, M., Kocher, J. A., and Eckel-Passow, J. E. A new statistic for identifying batch effects in high-throughput genomic data that uses guided principal components analysis. Bioinformatics, (in review).
}
\author{
Sarah Reese \email{reesese@vcu.edu}
}
\seealso{
\code{\link{gPCA.batchdetect}},\code{\link{PCplot}},\code{\link{CumulativeVarPlot}}
}
\examples{
# gDist(out)
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
